backend/requirements.txt
fastapi
pydantic>=2.0
motor
uvicorn
pytest
passlib[bcrypt]
python-multipart
PyJWT